Numpy and tensorflow conversion cleared and refactored.

+++ LIDCArtifactReduction/math_helpers/tensorflow_math_mixin.py
@@ -21,7 +21,11 @@ class TensorflowMathMixin(MathMixin):
         return tf.random.poisson([], mean_tf, dtype=tf.float32)
 
     def as_array(self, x):
-        return tf.cast(tf.convert_to_tensor(x), dtype=tf.float32)
+        return tf.convert_to_tensor(x, dtype=tf.float32)
+        # The row below generally would accept x to be Tensor with dtype=tf.float64 and cast it.
+        # tf.convert_to_tensor with dtype=tf.float32 does not accept such input but it is better to have this option,
+        # because tf.float64 should not be present in the system.
+        # return tf.cast(tf.convert_to_tensor(x), dtype=tf.float32)
 
     def where(self, condition, x, y):
         return tf.where(condition, x, y)
